backend/pyproject.toml
[project]
name = "trove"
version = "0.1.0"
description = "A modern replacement for FlexGet with multi-indexer search, Usenet support, and optional local AI."
readme = "../README.md"
requires-python = ">=3.12"
license = { text = "GPL-3.0-or-later" }
authors = [{ name = "Trove contributors" }]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlmodel>=0.0.22",
    "alembic>=1.14",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "httpx>=0.27",
    "apscheduler>=3.10",
    "structlog>=24.4",
    "passlib[argon2]>=1.7.4",
    "argon2-cffi>=23.1",
    "itsdangerous>=2.2",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",
    "guessit>=3.8",
    "parse-torrent-name>=1.1.1",
    "cryptography>=43.0",
    "litellm>=1.52",
    "feedparser>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "respx>=0.21",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-pyyaml",
]

[project.scripts]
trove = "trove.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trove"]

[tool.hatch.build.targets.wheel.force-include]
"src/trove/docs" = "trove/docs"

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",  # pycodestyle + pyflakes
    "I",            # isort
    "B",            # flake8-bugbear
    "UP",           # pyupgrade
    "N",            # pep8-naming
    "SIM",          # flake8-simplify
    "RUF",
]
ignore = [
    "E501",  # line length handled by formatter
    "B008",  # FastAPI idiom: `Depends(...)` in default args
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "src"
packages = ["trove"]
